Feature virtual dom and dom reducers
