+++ blocks/grid/grid.js
@@ -1,16 +1,14 @@
 import ComponentBase from '../../scripts/component-base.js';
-import { flat } from '../../scripts/libs.js';
+import { flat, stringToJsVal } from '../../scripts/libs.js';
 
 export default class Grid extends ComponentBase {
-  nestedComponentsConfig = {};
-
-  gridElements = [];
-
-  gridItemsElements = [];
+  // only one attribute is observed rest is set as css variables directly
+  static observedAttributes = ['data-reverse'];
 
   attributesValues = {
     all: {
       grid: {
+        // set props directly as css variables instead of data-attributes
         template: {
           columns: '1fr 1fr 1fr',
           rows: '1fr 1fr',
@@ -20,6 +18,9 @@ export default class Grid extends ComponentBase {
     },
   };
 
+  // use `grid` item as action from component base and apply those as css variables
+  // dinamic from {@link ../scripts/component-base.js:runConfigsByViewports}
+  // EG ${viewport}-grid-${attr}-"${value}"
   applyGrid(grid) {
     const f = flat(grid);
     Object.keys(f).forEach((key) => {
@@ -27,49 +28,46 @@ export default class Grid extends ComponentBase {
     });
   }
 
-  async connected() {
-    await this.collectGridItemsFromBlocks();
-  }
+  async onAttributeReverseChanged({ oldValue, newValue }) {
+    // await for initialization because access to this.gridItems is required;
+    await this.initialization;
 
-  async collectGridItemsFromBlocks() {
-    await this.checkIndexes(this.nextElementSibling);
-  }
+    if (oldValue === newValue) return;
 
-  async checkIndexes() {
-    const siblings = Array.from(this.parentNode.children);
-    // check index of this element
-    this.dataset.index = siblings.indexOf(this);
-    // verify other grid elements
-    this.gridElements = Array.from(this.parentNode.querySelectorAll('raqn-grid')).filter((grid) => {
-      grid.dataset.index = siblings.indexOf(grid);
-      return grid.dataset.index > this.dataset.index;
+    const val = stringToJsVal(newValue);
+    const reverse = val === true || newValue === 'alternate';
+
+    let items = [...this.children];
+
+    switch (val) {
+      case true:
+        items = [...this.children].reverse();
+        break;
+      case 'alternate':
+        items = this.alternateReverse();
+        break;
+      default:
+        break;
+    }
+
+    items.forEach((item, index) => {
+      if (reverse) {
+        item.dataset.order = index + 1;
+      } else {
+        delete item.dataset.order;
+      }
     });
-    // get max index of the next grid item
-    const nextGrid = this.gridElements.length > 0 ? siblings.indexOf(this.gridElements[0]) : siblings.length;
-    // get all grid items between this and next grid
-    this.gridItemsElements = Array.from(this.parentNode.querySelectorAll('.grid-item')).filter((item) => {
-      item.dataset.index = siblings.indexOf(item);
-      return item.dataset.index > siblings.indexOf(this) && item.dataset.index <= nextGrid;
-    });
-
-    let previous = siblings.indexOf(this) + 1;
-
-    return Promise.allSettled(
-      this.gridItemsElements.map(async (item) => {
-        const children = Array.from(this.parentNode.children).slice(previous, item.dataset.index - 1);
-        const configByClasses = [...item.classList];
-        previous = item.dataset.index;
-        item.remove();
-        return this.createGridItem(children, configByClasses);
-      }),
-    );
   }
 
-  async createGridItem(children, configByClasses) {
-    const tempGridItem = document.createElement('raqn-grid-item');
-    tempGridItem.init({ configByClasses });
-    tempGridItem.append(...children);
-    await this.append(tempGridItem);
-    return tempGridItem;
+  // swaps every 2 items [1,2,3,4,5,6,7] => [2,1,4,3,6,5,7]
+  alternateReverse() {
+    return [...this.children].reduce((acc, x, i, arr) => {
+      if ((i + 1) % 2) {
+        if (arr.length === i + 1) acc.push(x);
+        return acc;
+      }
+      acc.push(x, arr[i - 1]);
+      return acc;
+    }, []);
   }
 }
